pyproject.toml
[project]
name = "V1Model_Simulation"
version = "0.1.0"
description = "Simulation of mouse V1 model"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bctpy>=0.6.1",
    "matplotlib>=3.10.8",
    "networkx>=3.6.1",
    "numpy>=2.4.3",
    "python-louvain>=0.16",
    "scipy>=1.17.1",
    "typer>=0.24.1",
]
